+++ alphastats/plots/clustermap.py
@@ -6,7 +6,7 @@ import seaborn as sns
 
 from alphastats.DataSet_Statistics import Statistics
 from alphastats.keys import Cols
-from alphastats.plots.PlotUtils import PlotUtils
+from alphastats.plots.plot_utils import PlotUtils
 
 
 class ClusterMap(PlotUtils):
