+++ alphastats/plots/dimensionality_reduction.py
@@ -8,7 +8,7 @@ import sklearn
 
 from alphastats.DataSet_Preprocess import Preprocess
 from alphastats.keys import Cols
-from alphastats.plots.PlotUtils import PlotlyObject, PlotUtils
+from alphastats.plots.plot_utils import PlotlyObject, PlotUtils
 
 # make own alphastats theme
 plotly.io.templates["alphastats_colors"] = plotly.graph_objects.layout.Template(
