+++ alphastats/plots/volcano_plot.py
@@ -10,7 +10,7 @@ from alphastats.DataSet_Preprocess import PreprocessingStateKeys
 from alphastats.DataSet_Statistics import Statistics
 from alphastats.keys import Cols
 from alphastats.multicova import multicova
-from alphastats.plots.PlotUtils import PlotlyObject, PlotUtils
+from alphastats.plots.plot_utils import PlotlyObject, PlotUtils
 from alphastats.statistics.DifferentialExpressionAnalysis import (
     DifferentialExpressionAnalysis,
 )
